updates to add activity

+++ app/src/main/java/com/example/aurorasheetapp/MainActivity.java
@@ -238,7 +238,7 @@ public class MainActivity extends AppCompatActivity implements
                     make,
                     Integer.parseInt(serial),
                     model,
-                    Integer.parseInt(value),
+                    Double.parseDouble(value),
                     comment
             );
             listItems.add(listItem);
